requirements.txt
llama-index==0.11.3
llama-index-vector-stores-lancedb==0.2.4
llama-index-embeddings-huggingface==0.3.1
huggingface_hub==0.26.1
lancedb==0.24.0
pytubefix==6.16.3
moviepy
faster-whisper==1.0.3
pillow==10.4.0
pytesseract
sentence-transformers==3.0.1
transformers==4.45.1
torch==2.4.1
google-generativeai==0.8.3
pyspellchecker==0.8.1
spellchecker
fastapi
uvicorn
python-dotenv==1.0.1
huggingface_hub==0.26.1
matplotlib==3.9.2